user distance calculation

+++ app/src/main/java/com/pandacat/simplechaserun/data/states/MonsterState.kt
@@ -1,6 +1,6 @@
 package com.pandacat.simplechaserun.data.states
 
-data class MonsterState(val activeState: State, val distanceM: Int, val activeTimeMillis: Long) {
+data class MonsterState(val activeState: State, val distanceM: Double, val activeTimeMillis: Long) {
     enum class State
     {
         NOT_STARTED,
