+++ app/src/main/java/com/pandacat/simplechaserun/data/states/RunnerState.kt
@@ -3,5 +3,5 @@ package com.pandacat.simplechaserun.data.states
 import com.google.android.gms.maps.model.LatLng
 
 data class RunnerState(val currentPosition: LatLng,
-                       val totalDistanceM: Int,
+                       val totalDistanceM: Double,
                        val totalTimeMillis: Long)
